Detection of Forest fire using MATLAB
A land based multi spectral and multi temporal video has been proposed to determine whether smoke and fire is present in the scene. This method will use temporal, spatial and spectral information from multiple camera views to identify the smoke. The special feature of the smoke is compared to the special feature of the other objects in the same to identify the smoke’s spectral signature. Using various images through time, the temporal variance of the smoke will be identified with the simple difference in the image and principal component transportation, this is a complicated method which provides better results.