Detection of bone fracture using MAT lab
The bone fracture is a very common problem in human beings which occurs due to high
pressure applied on bones and also sometimes due to osteoporosis and bone cancer. Hence the
accurate the diagnosis of the bone fracture is very important aspect . Generally we use Xrays
and CT scan images to detect the bone fracture. The aim of the project is to develop an image
processing technique for a quick and accurate classification of bone fracture based on the
information gained from the Xray nd CT scan images. The images obtained are been processed
through various techniques like pre-processing, segmentation, edge detection and feature
extraction method. Then the processed images are further classified into fractured and non-
fractured bone and compare the accuracy of different methods. This project is fully been
employed by MATLAB as the programming tool for loading and image processing and user
interface development. Results obtained will demonstrate the performance of the bone
fracture detection system with some limitations and good accuracy.